Attr eng: float vs int trouble solved
git-svn-id: https://svn.opendtect.org/od/trunk@38544 a985855d-d384-4b45-ae33-31a410998c3c

+++ src/AttributeEngine/attribengman.cc
@@ -369,7 +369,7 @@ const DataCubes* EngineMan::getDataCubesOutput( const Processor& proc )
 			if ( Values::isUdf( val ) )
 			    continue;
 
-			const int outzsampidx = Math::Floor( outfidx );
+			const int outzsampidx = (int)Math::Floor( outfidx );
 			output->setValue( cubeidx, inlidx, crlidx, outzsampidx, val );
 		    }
 		}
